backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
pydantic-settings>=2.2
python-dotenv>=1.0
email-validator>=2.1
